refactor: split out WIP pact publication methods into their own module
